backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
google-auth
psycopg2-binary
pydantic>=2.0
pydantic-settings>=2.0
alembic
requests
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
jinja2
apscheduler
accelerate
langchain
redis
transformers
protobuf
faiss-gpu
openai
pytest
pytest-asyncio
httpx
black
isort
flake8
mypy
bitsandbytes
optimum
langchain-community
sentence-transformers
google-generativeai
langchain-google-genai
auto-gptq

--extra-index-url https://download.pytorch.org/whl/cu126

torch
torchvision
torchaudio

